fix: typescript screaming when using ActionAggregator
- more UTs
