+++ src/types.ts
@@ -34,7 +34,7 @@ export type StrictSchema<Target = any, Source = any> = {
   [destinationProperty in keyof Target]:
     | ActionString<Source>
     | ActionFunction<Target, Source, Target[destinationProperty]>
-    | ActionAggregator
+    | ActionAggregator<Source>
     | ActionSelector<Source, Target[destinationProperty]>
     | StrictSchema<Target[destinationProperty], Source>
 };
@@ -43,7 +43,7 @@ export type Schema<Target = any, Source = any> = {
   [destinationProperty in keyof Target]?:
     | ActionString<Source>
     | ActionFunction<Target, Source, Target[destinationProperty]>
-    | ActionAggregator
+    | ActionAggregator<Source>
     | ActionSelector<Source, Target[destinationProperty]>
     | Schema<Target[destinationProperty], Source>
 };
@@ -106,7 +106,8 @@ export interface ActionFunction<D = any, S = any, R = any> {
  * ```
  *
  */
-export type ActionString<T> = keyof T;
+export type ActionString<Source> = string; // TODO: ActionString should support string and string[] for deep properties
+
 /**
  * An Array of String that allows to perform a function over source property
  *
@@ -125,7 +126,7 @@ export type ActionString<T> = keyof T;
  * //=> { fooAndBar: { foo: 'foo', bar: 'bar' } }
  * ```
  */
-export type ActionAggregator = string[];
+export type ActionAggregator<T extends unknown = unknown> = T extends object ? (keyof T)[] | string[] : string[];
 /**
  * An Object that allows to perform a function over a source property's value
  *
@@ -149,10 +150,10 @@ export type ActionAggregator = string[];
  *```
  *
  */
-export type ActionSelector<Source = any, R = any> = {
-  path: string | string[];
+export interface ActionSelector<Source = object, R = any> {
+  path: ActionString<Source> | ActionAggregator<Source>;
   fn: (fieldValue: any, object: Source, items: Source, objectToCompute: R) => R;
-};
+}
 
 export interface Constructable<T> {
   new (...args: any[]): T;
